Response payload format changes.
Modified the way the payload is returned to allow for a status and error reporting.
Quickly wrote a manual filter for the ObjectID _id's as the ObjectID is not serializable
and json.dumps() freaks out over it. Also made a change to the pymongo model for host
home hostnames are stored (this is make it work more smoothly with AngularJS)